backend/requirements.txt
backend==0.2.4.1
fastapi==0.115.4
pydantic==2.9.2
SQLAlchemy==2.0.31
authlib==1.3.2
python-jose==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
email-validator==2.2.0
bcrypt==4.2.0
